Mover interface dos controllers para dentro do pacote Core
