+++ src/core/SnackTech.Core/Interfaces/IClienteController.cs
@@ -1,0 +1,10 @@
+using SnackTech.Common.Dto;
+
+namespace SnackTech.Core.Interfaces;
+
+public interface IClienteController
+{
+    Task<ResultadoOperacao<ClienteDto>> CadastrarNovoCliente(ClienteSemIdDto clienteSemIdDto);
+    Task<ResultadoOperacao<ClienteDto>> IdentificarPorCpf(string cpf);
+    Task<ResultadoOperacao<ClienteDto>> SelecionarClientePadrao();
+}
