+++ src/core/SnackTech.Core/Interfaces/IPedidoController.cs
@@ -1,0 +1,8 @@
+using SnackTech.Common.Dto;
+
+namespace SnackTech.Core.Interfaces;
+
+public interface IPedidoController
+{
+    Task<ResultadoOperacao<Guid>> IniciarPedido(string? cpfCliente);
+}
