+++ src/core/SnackTech.Core/Interfaces/IProdutoController.cs
@@ -1,0 +1,11 @@
+using SnackTech.Common.Dto;
+
+namespace SnackTech.Core.Interfaces;
+
+public interface IProdutoController
+{
+    Task<ResultadoOperacao<ProdutoDto>> CadastrarNovoProduto(ProdutoSemIdDto produtoSemIdDto);
+    Task<ResultadoOperacao<ProdutoDto>> EditarProduto(Guid identificacao, ProdutoSemIdDto produtoParaAlterar);
+    Task<ResultadoOperacao> RemoverProduto(Guid id);
+    Task<ResultadoOperacao<IEnumerable<ProdutoDto>>> BuscarProdutosPorCategoria(int categoriaId);
+}
